[kernel] Last compiler warning cleanup for kernel and libc, other cleanup

+++ elks/arch/i86/drivers/char/ntty.c
@@ -216,7 +216,7 @@ void tty_echo(register struct tty *tty, unsigned char ch)
  *
  */
 
-size_t tty_write(struct inode *inode, struct file *file, char *data, int len)
+size_t tty_write(struct inode *inode, struct file *file, char *data, size_t len)
 {
     register struct tty *tty = determine_tty(inode->i_rdev);
     int i, s;
@@ -236,7 +236,7 @@ size_t tty_write(struct inode *inode, struct file *file, char *data, int len)
     return i;
 }
 
-size_t tty_read(struct inode *inode, struct file *file, char *data, int len)
+size_t tty_read(struct inode *inode, struct file *file, char *data, size_t len)
 {
     register struct tty *tty = determine_tty(inode->i_rdev);
     int i = 0;
